+++ app/assets/javascripts/folio/devise.js
@@ -1,0 +1,2 @@
+//= require folio/devise/modal/modal
+//= require folio/devise/omniauth/forms/forms