+++ app/cells/folio/devise/omniauth/forms/forms.js
@@ -1,0 +1,5 @@
+$(document)
+  .on('click', '.f-js-devise-omniauth-forms-trigger', (e) => {
+    e.preventDefault()
+    $(`.f-devise-omniauth-forms__button[data-provider="${e.currentTarget.dataset.provider}"]`).click()
+  })