+++ app/cells/folio/devise/omniauth/forms_cell.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+class Folio::Devise::Omniauth::FormsCell < Folio::Devise::ApplicationCell
+  TRIGGER_CLASS_NAME = "f-js-devise-omniauth-forms-trigger"
+
+  def show
+    render if ::Rails.application.config.folio_users_omniauth_providers.present?
+  end
+
+  def omniauth_sign_in_form(key, &block)
+    url = controller.main_app.send("user_#{key}_omniauth_authorize_path")
+    simple_form_for "", url:, &block
+  end
+end