+++ app/cells/folio/devise/omniauth/icon_cell.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class Folio::Devise::Omniauth::IconCell < Folio::ApplicationCell
+class Folio::Devise::Omniauth::IconCell < Folio::Devise::ApplicationCell
   def size_class_name
     if options[:size]
       "f-devise-omniauth-icon--size-#{options[:size]}"