+++ lib/folio/engine.rb
@@ -54,7 +54,7 @@ module Folio
     config.folio_users_confirmable = false
     config.folio_users_publicly_invitable = true
     config.folio_users_use_address = true
-    config.folio_users_omniauth_providers = %i[facebook google_oauth2 twitter]
+    config.folio_users_omniauth_providers = %i[facebook google_oauth2 twitter2]
     config.folio_users_after_ajax_sign_up_redirect = false
     config.folio_users_after_sign_in_path = :root_path
     config.folio_users_after_sign_up_path = :root_path