+++ test/cells/folio/devise/omniauth/icon_cell_test.rb
@@ -7,7 +7,7 @@ class Folio::Devise::Omniauth::IconCellTest < Cell::TestCase
     html = cell("folio/devise/omniauth/icon", nil).(:show)
     assert html.has_css?(".f-devise-omniauth-icon")
 
-    html = cell("folio/devise/omniauth/icon", :twitter).(:show)
+    html = cell("folio/devise/omniauth/icon", :twitter2).(:show)
     assert html.has_css?(".f-devise-omniauth-icon")
   end
 end